# Refactor, Code Clean up Todo Sample Controller Removed #Removed
Refactor, Code Clean up #Removed SampleCodes from LightBend
- [x] Todo Sample Controller Removed.
- [x] V1 Sample Controllers Removed.
- [x] Models Removed.
