+++ app/controllers/controllerRoutes/CampaignApiRouter.scala
@@ -1,6 +1,7 @@
 package controllers.controllerRoutes
 
 import controllers.apis.CampaignsApiController
+import controllers.controllerRoutes.routerGeneric.AbstractGenericRouter
 import javax.inject.Inject
 import shared.com.ortb.persistent.schema.Tables._
