--- app/models/Widget.scala
@@ -1,10 +1,0 @@
-package models
-
-/**
- * Presentation object used for displaying data in a template.
- *
- * Note that it's a good practice to keep the presentation DTO,
- * which are used for reads, distinct from the form processing DTO,
- * which are used for writes.
- */
-case class Widget(name: String, price: Int)
