--- app/v1/post/PostRouter.scala
@@ -1,30 +1,0 @@
-package v1.post
-
-import javax.inject.Inject
-import play.api.routing.Router.Routes
-import play.api.routing.SimpleRouter
-import play.api.routing.sird._
-
-/**
-  * Routes and URLs to the PostResource controller.
-  */
-class PostRouter @Inject()(controller: PostController) extends SimpleRouter {
-  val prefix = "/v1/posts"
-
-  def link(id: PostId): String = {
-    import io.lemonlabs.uri.dsl._
-    val url = prefix / id.toString
-    url.toString()
-  }
-
-  override def routes: Routes = {
-    case GET(p"/") =>
-      controller.index
-
-    case POST(p"/") =>
-      controller.process
-
-    case GET(p"/$id") =>
-      controller.show(id)
-  }
-}
